Split admin pricing page
